MobileDock: Fix saving to file, fix forced jettison, various fixes

+++ Plugins/Public/mobiledocking_plugin/Main.h
@@ -52,7 +52,7 @@ void SendSetBaseInfoText2(UINT client, const wstring &message);
 void ForceLaunch(uint client);
 
 // The distance to undock from the carrier
-static int set_iMobileDockOffset = 100;
+static float set_iMobileDockOffset = 100.0f;
 
 // A map of all docking requests pending approval by the carrier
 // key = docking client id, value = carrier client id
